Create 01-aesix
This file must be in "/etc/update-motd.d/" must be executable, and must be the only executable script running on ssh login for aesthetic purposes.  Additional work will be needed to make it work with multiple IPs.  I only have 1 WAN routable IP (1 each, v4 and v6) and as such, this script works perfectly fine for my needs.

This script requires an additional file to be located at "~/ssh-motd" for which the user can alter.  
Alternatively, the ssh-motd file can be hosted in a non-user area, the script re-directed to it.  
This file is used by me for the welcome section.  If using a global file, the $USERNAME variable and scripts can be commented out.